requirements.txt
# Core FastAPI and server
fastapi
uvicorn[standard]
pydantic
pydantic-settings

# Async HTTP client
httpx

# Redis for caching
redis[hiredis]

# LangChain and LLMs
langchain 
langchain-text-splitters 
langchain-community
langchain[google-genai]
langgraph
langsmith

# Vector database
chromadb

# Data processing
pandas
numpy

# ML libraries
scikit-learn
xgboost

# Medical imaging
tensorflow
Pillow

# FHIR client
fhirclient==4.2.1

# Reading level analysis
textstat

# Environment variables
python-dotenv

"mcp[cli]" fastmcp